server/python/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
flake8==4.0.1
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
requests==2.27.1
stripe==2.67.0
urllib3==1.26.8
Werkzeug==2.0.3
yapf==0.32.0
